Delete jax arrays as we go, as they are not automatically freed.
Delete jax arrays as we go, as they are not automatically freed when
going out of scope. On CPU, garbage collection eventually get them,
but for GPU, this just causes the program to run out of GPU memory
and crash.

With explicit del's, the volume matcher now runs on GPU.