Added metadata to 0700 and fixed a small typo
Metadata for the full array of coordinates for the center of cylinder
have been added. These are used for radial masking of the segmented
volumes and the ground truth when doing the final evaluation. Also the
radius is added, even though the attribute "implant_radius" already
exists. This however uses the Cs and Rs variables, which gives a slight
numerical difference, probably just from rounding. We might want to
delete this other attribute at some point.