added novisim as metadata flag for h5 files
For h5 files that are not related to novisim, the attribute "novisim" will
have value 0, or otherwise 1. This avoids checks being hardcoding during
the processing steps.

There also seems to be an error in the string recieved for
meta["ByteOrder"] which contains a whitespace and a semicolon. For now
the quick fix was to strip the string and keep only the first part.